Refactor: 객체 인식 역할 코드 별도 파일로의 분리 #65
